# openapi.yaml
openapi: 3.0.1
info:
  title: Document Management System API
  description: |
    A comprehensive Document Management System REST API with features including:

    - **Document Management**: Create, read, update, and delete documents
    - **Content Management**: Upload and download content with flexible storage options
    - **Versioning**: Major and minor version control for documents
    - **Content Renditions**: Primary and secondary content renditions with transformations
    - **Full-Text Search**: Lucene-powered search with fielded queries
    - **File Store Management**: Manage multiple file storage locations
    - **Content Transformations**: PDF-to-text and other content transformations

    ## Key Features

    ### Document Types
    - ARTICLE, MANUAL, REPORT, SPREADSHEET, PRESENTATION, IMAGE, VIDEO, AUDIO, OTHER

    ### Storage Options
    - Database storage for small files
    - File system storage for large files
    - Automatic migration between storage types

    ### Search Capabilities
    - Simple text search across all fields
    - Fielded search (name, description, keywords, tags, content)
    - Boolean operators (AND, OR)
    - Phrase search, wildcard search
    - Indexes all document metadata and indexable content

    ### Content Renditions
    - Primary renditions (original content)
    - Secondary renditions (transformed content)
    - Automatic text extraction from PDFs
    - Extensible transformer plugin framework
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/documents:
    get:
      tags:
      - Documents
      summary: Get all items
      description: Retrieve all items including all versions
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
    post:
      tags:
      - Documents
      summary: Create item
      description: Create a new item
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
        required: true
      responses:
        "400":
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        "201":
          description: Item created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
...
  /api/search:
    get:
      tags:
      - Search
      summary: Simple text search
      description: "Search across all document fields (name, description, keywords,\
        \ tags, content) and indexable content using Lucene query syntax. Supports\
        \ wildcards (*), phrase queries (\"exact phrase\"), and boolean operators\
        \ (AND, OR, NOT)."
      operationId: search
      parameters:
      - name: q
        in: query
        description: Search query using Lucene syntax
        required: true
        schema:
          type: string
        example: spring framework
      - name: limit
        in: query
        description: Maximum number of results to return
        required: false
        schema:
          type: integer
          format: int32
          default: 50
        example: 50
      responses:
        "200":
          description: Search completed successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        "400":
          description: Invalid query syntax
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
...
  /api/documents/search:
    get:
      tags:
      - Documents
      summary: Search documents by keywords
      description: Search for documents using keyword matching (searches in the keywords
        field)
      operationId: searchByKeywords
      parameters:
      - name: keywords
        in: query
        description: Keywords to search for
        required: true
        schema:
          type: string
        example: spring boot java
      responses:
        "200":
          description: Documents found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
...
  /api/documents/by-type/{documentType}:
    get:
      tags:
      - Documents
      summary: Find documents by type
      description: "Retrieve all documents of a specific type (ARTICLE, MANUAL, REPORT,\
        \ etc.)"
      operationId: findByDocumentType
      parameters:
      - name: documentType
        in: path
        description: Document type to search for
        required: true
        schema:
          type: string
          enum:
          - ARTICLE
          - REPORT
          - CONTRACT
          - MANUAL
          - PRESENTATION
          - TRIP_REPORT
          - OTHER
      responses:
        "200":
          description: Documents found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
  /api/documents/by-tag/{tag}:
    get:
      tags:
      - Documents
      summary: Find documents by tag
      description: Retrieve all documents that contain the specified tag
      operationId: findByTag
      parameters:
      - name: tag
        in: path
        description: Tag to search for
        required: true
        schema:
          type: string
        example: java
      responses:
        "200":
          description: Documents found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentDTO'
...
components:
  schemas:
...
    ContentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contentType:
          type: string
        sysObjectId:
          type: integer
          format: int64
        fileStoreId:
          type: integer
          format: int64
        fileStoreName:
          type: string
        storagePath:
          type: string
        size:
          type: integer
          format: int64
        formattedSize:
          type: string
        storageType:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    DocumentDTO:
      required:
      - documentType
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        majorVersion:
          type: integer
          format: int32
        minorVersion:
          type: integer
          format: int32
        parentVersionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        contents:
          type: array
          items:
            $ref: '#/components/schemas/ContentDTO'
        description:
          type: string
        documentType:
          type: string
          enum:
          - ARTICLE
          - REPORT
          - CONTRACT
          - MANUAL
          - PRESENTATION
          - TRIP_REPORT
          - OTHER
        tags:
          uniqueItems: true
          type: array
          items:
            type: string
        keywords:
          type: string
    SearchResult:
      type: object
      properties:
        documentId:
          type: integer
          format: int64
        score:
          type: number
          format: float
        name:
          type: string
        description:
          type: string
        keywords:
          type: string
        tags:
          type: string
...